pyproject.toml
[tool.poetry]
name = "newsful-streamlit"
version = "0.1.0"
description = ""
authors = ["Siddhesh Agarwal <siddhesh.agarwal@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.39.0"
instructor = "^1.6.3"
openai = "^1.53.0"
ujson = "^5.10.0"
pydantic = "^2.9.2"
google-generativeai = "^0.8.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
